+++ CHANGES.md
@@ -6,6 +6,7 @@ This file includes at least a partial list of the major changes in each version.
 v0.3.1 (in progress)
 --------------------
 
+- added `Sequence#apply`
 - added `ObjectLikeSequence#merge`
 - added `Sequence#toStream` to create an actual `stream.Readable` in Node, to interact w/ stream-based libraries